+++ pkgs/test/lib/src/runner/browser/compilers/precompiled.dart
@@ -12,7 +12,9 @@ import 'package:shelf/shelf_io.dart' as shelf_io;
 import 'package:shelf_packages_handler/shelf_packages_handler.dart';
 import 'package:shelf_static/shelf_static.dart';
 import 'package:shelf_web_socket/shelf_web_socket.dart';
-import 'package:test_api/backend.dart' show StackTraceMapper, SuitePlatform;
+import 'package:test_api/backend.dart'
+    show Compiler, StackTraceMapper, SuitePlatform;
+import 'package:test_core/src/runner/configuration.dart'; // ignore: implementation_imports
 import 'package:test_core/src/runner/suite.dart'; // ignore: implementation_imports
 import 'package:test_core/src/util/package_config.dart'; // ignore: implementation_imports
 import 'package:test_core/src/util/stack_trace_mapper.dart'; // ignore: implementation_imports
@@ -24,16 +26,16 @@ import '../../../util/package_map.dart';
 import '../../../util/path_handler.dart';
 import 'compiler_support.dart';
 
+class JsPrecompiledSupport = PrecompiledSupport with JsHtmlWrapper;
+class WasmPrecompiledSupport = PrecompiledSupport with WasmHtmlWrapper;
+
 /// Support for precompiled test files.
-class PrecompiledSupport implements CompilerSupport {
+abstract class PrecompiledSupport extends CompilerSupport {
   /// Whether [close] has been called.
   bool _closed = false;
 
   /// Mappers for Dartifying stack traces, indexed by test path.
   final _mappers = <String, StackTraceMapper>{};
-
-  /// A [PathHandler] used to serve test specific artifacts.
-  final _pathHandler = PathHandler();
 
   /// The root directory served statically by the server.
   final String _root;
@@ -60,12 +62,16 @@ class PrecompiledSupport implements CompilerSupport {
   @override
   Uri get serverUrl => _server.url.resolve('$_secret/');
 
-  PrecompiledSupport._(this._server, this._root, String faviconPath) {
+  PrecompiledSupport._(super.config, super.defaultTemplatePath, this._server,
+      this._root, String faviconPath) {
     var cascade = shelf.Cascade()
         .add(_webSocketHandler.handler)
+        .add(createStaticHandler(_root, serveFilesOutsidePath: true))
+        // TODO: This packages dir handler should not be necessary?
         .add(packagesDirHandler())
-        .add(_pathHandler.handler)
-        .add(createStaticHandler(_root));
+        // Even for precompiled tests, we will auto-create a bootstrap html file
+        // if none was present.
+        .add(htmlWrapperHandler);
 
     var pipeline = const shelf.Pipeline()
         .addMiddleware(PathHandler.nestedIn(_secret))
@@ -78,11 +84,25 @@ class PrecompiledSupport implements CompilerSupport {
   }
 
   static Future<PrecompiledSupport> start({
+    required Compiler compiler,
+    required Configuration config,
+    required String defaultTemplatePath,
     required String root,
     required String faviconPath,
   }) async {
     var server = shelf_io.IOServer(await HttpMultiServer.loopback(0));
-    return PrecompiledSupport._(server, root, faviconPath);
+
+    return switch (compiler) {
+      Compiler.dart2js => JsPrecompiledSupport._(
+          config, defaultTemplatePath, server, root, faviconPath),
+      Compiler.dart2wasm => WasmPrecompiledSupport._(
+          config, defaultTemplatePath, server, root, faviconPath),
+      Compiler.exe ||
+      Compiler.kernel ||
+      Compiler.source =>
+        throw UnsupportedError(
+            'The browser platform does not support $compiler'),
+    };
   }
 
   /// Compiles [dartPath] using [suiteConfig] for [platform].